定期rebootを一度設定するとoffできない問題の対応

+++ atomcam_configs/overlay_rootfs/scripts/reboot.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
-sync
-sync
-sync
-reboot
+HACK_INI=/media/mmc/hack.ini
+REBOOT=$(awk -F "=" '/REBOOT *=/ {print $2}' $HACK_INI)
+
+if [ "$REBOOT" = "on" ]; then
+  sync
+  sync
+  sync
+  reboot
+fi
